requirements.txt
requests
pandas
dotenv
pyodbc
json
os
sqlalchemy
urllib
sqlalchemy
utilities
azure-cognitiveservices-vision-computervision
unidecode 
streamlit
